+++ iextrading4j-client/src/main/java/pl/zankowski/iextrading4j/client/rest/request/stocks/v1/AbstractLastStocksRequestBuilder.java
@@ -23,9 +23,9 @@ public abstract class AbstractLastStocksRequestBuilder<R, B extends IRestRequest
         return (B) this;
     }
 
-    abstract RestRequest<R> request();
+    protected abstract RestRequest<R> request();
 
-    abstract RestRequest<R> requestWithLast();
+    protected abstract RestRequest<R> requestWithLast();
 
     @Override
     public RestRequest<R> build() {