+++ iextrading4j-client/src/test/java/pl/zankowski/iextrading4j/client/rest/request/stocks/EarningsRequestBuilderTest.java
@@ -4,6 +4,7 @@ import org.junit.Test;
 import pl.zankowski.iextrading4j.api.stocks.Earnings;
 import pl.zankowski.iextrading4j.client.rest.manager.MethodType;
 import pl.zankowski.iextrading4j.client.rest.manager.RestRequest;
+import pl.zankowski.iextrading4j.client.rest.request.stocks.v1.Period;
 
 import javax.ws.rs.core.GenericType;
 
@@ -25,6 +26,25 @@ public class EarningsRequestBuilderTest {
         assertThat(request.getResponseType()).isEqualTo(new GenericType<Earnings>() {});
         assertThat(request.getPathParams()).containsExactly(entry("symbol", symbol));
         assertThat(request.getQueryParams()).isEmpty();
+    }
+
+    @Test
+    public void shouldSuccessfullyCreateRequestWithLastAndPeriod() {
+        final String symbol = "IBM";
+        final Period period = Period.ANNUAL;
+        final int last = 3;
+
+        final RestRequest<Earnings> request = new EarningsRequestBuilder()
+                .withSymbol(symbol)
+                .withLast(last)
+                .withPeriod(period)
+                .build();
+
+        assertThat(request.getMethodType()).isEqualTo(MethodType.GET);
+        assertThat(request.getPath()).isEqualTo("/stock/{symbol}/earnings");
+        assertThat(request.getResponseType()).isEqualTo(new GenericType<Earnings>() {});
+        assertThat(request.getPathParams()).contains(entry("symbol", symbol), entry("last", String.valueOf(last)));
+        assertThat(request.getQueryParams()).contains(entry("period", period.name().toLowerCase()));
     }
 
 }